User commit

+++ src/main/java/com/example/demo/service/ValidationService.java
@@ -4,5 +4,5 @@ import com.example.demo.entity.ValidationEntity;
 import java.util.List;
 public interface ValidationService{
     ValidationEntity postValid(ValidationEntity val);
-    ValidationEntity findData(int id);
+    ValidationEntity findData(Long id);
 }
